+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All Notable changes to `StadGent/drupal_theme_gent-base`.
 - SGD8-3000: Fixed bug with Drupal behaviors breaking main menu on pages with tables.
 - SGD8-3000: Fixed bug with colIndex push.
 - HEAT-6832177: Fix cache key for teaser_with_links.
+- SGD8-2983: Fix facets range slider missing range values.
 
 ## [6.0.4]